fix: add use client directive to page.tsx and wrap useSearchParams in suspense boundary
