+++ src/components/index.ts
@@ -4,7 +4,8 @@ import Footer from "./Footer";
 import Divider from "./Divider";
 import Icon from "./Icon";
 import ChatBubble from "./ChatBubble";
+import ScrollToSection from "./ScrollToSection";
 
-export { Header, Button, Footer, Divider, Icon, ChatBubble };
+export { Header, Button, Footer, Divider, Icon, ChatBubble, ScrollToSection };
 export { default as ServiceCard } from "./ServiceCard";
 export { default as ProjectCard } from "./ProjectCard";
