+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 import useResponsiveType, { ResponsiveType } from "./useResponsiveType";
-import useAppResponse from "./useAppResponse";
+
 import useDeviceType from "./useDeviceType";
-export { useResponsiveType, useAppResponse, useDeviceType };
+export { useResponsiveType, useDeviceType };
 export type { ResponsiveType };
