--- src/libs/axios/index.ts
@@ -1,5 +1,0 @@
-import * as auth from "./auth";
-
-const restAPI = { auth };
-
-export default restAPI;
